+++ src/messages/vi/admin.php
@@ -367,4 +367,19 @@ return [
     'js_pagination_page' => 'Page',
     'dashboard_log_message_delete' => 'Deleted a record in <strong>{container}</strong>.',
     'model_tag_relations_count' => 'Used',
+
+// 1.2.3
+
+    'active_button_duplicate_label' => 'Duplicate',
+    'active_button_duplicate_success' => 'A copy has been created.',
+    'active_button_duplicate_error' => 'Error while creating the copy: {message}',
+
+// 1.3
+
+    'model_tag_translation' => 'Translation',
+    'model_tag_translation_hint' => 'An optional translation for the tag in the given language. This allows to assign and output the same tag on multi-language websites.',
+    'active_button_timestamp_label' => 'Timestamp',
+    'active_button_timestamp_success' => 'Timestamp has been set successfull.',
+    'active_button_timestamp_error' => 'Error while storing timestamp.',
+
 ];