+++ src/resources/js/controllers.js
@@ -18,7 +18,8 @@
 	 *
 	 * + bool $config.inline Determines whether this crud is in inline mode orno
 	 */
-	zaa.controller("CrudController", ['$scope', '$rootScope', '$filter', '$http', '$sce', '$state', '$timeout', '$injector', '$q', 'AdminLangService', 'LuyaLoading', 'AdminToastService', 'CrudTabService', 'ServiceImagesData', function($scope, $rootScope, $filter, $http, $sce, $state, $timeout, $injector, $q, AdminLangService, LuyaLoading, AdminToastService, CrudTabService, ServiceImagesData) {
+	zaa.controller("CrudController", ['$scope', '$rootScope', '$filter', '$http', '$sce', '$state', '$timeout', '$injector', '$q', 'AdminLangService', 'AdminToastService', 'CrudTabService', 'ServiceImagesData', 
+	function($scope, $rootScope, $filter, $http, $sce, $state, $timeout, $injector, $q, AdminLangService, AdminToastService, CrudTabService, ServiceImagesData) {
 
 		$scope.toast = AdminToastService;
 
@@ -168,6 +169,36 @@
 			$scope.config.orderBy = sort + field;
 			$http.post('admin/api-admin-common/ngrest-order', {'apiEndpoint' : $scope.config.apiEndpoint, sort: sort, field: field}, { ignoreLoadingBar: true });
 			$scope.loadList();
+		};
+
+		/****************** ACTIVE BUTTON ***********/
+
+		$scope.callActiveButton = function(hash, id, event) {
+			var elmn = angular.element(event.currentTarget);
+			elmn.addClass('crud-buttons-button-loading');
+			$http.get($scope.config.apiEndpoint + '/active-button?hash=' + hash + '&id=' + id.join()).then(function(success) {
+				elmn.removeClass('crud-buttons-button-loading');
+				elmn.addClass('crud-buttons-button-success');
+				$timeout(function() {
+					elmn.removeClass('crud-buttons-button-success');
+				}, 5000);
+
+				angular.forEach(success.data.events, function(value) {
+					// event names
+					if (value == 'loadList') {
+						$scope.loadList();
+					}
+				});
+
+				AdminToastService.success(success.data.message);
+			}, function(error) {
+				elmn.removeClass('crud-buttons-button-loading');
+				elmn.addClass('crud-buttons-button-danger');
+				$timeout(function() {
+					elmn.removeClass('crud-buttons-button-danger');
+				}, 5000);
+				AdminToastService.error(error.data.message);
+			});
 		};
 
 		/***************** ACTIVE WINDOW *********/
@@ -324,6 +355,7 @@
 				$scope.loadList();
 				$scope.applySaveCallback();
 				$scope.switchTo(0, true);
+				$scope.resetData();
 			}, function(data) {
 				$scope.printErrors(data.data);
 			});
@@ -408,8 +440,8 @@
 			angular.forEach($scope.data.listArray, function(value, key) {
 				var json = {};
 				json[fieldName] = key;
-				var rowId = value[$scope.config.pk];
-				$http.put($scope.config.apiEndpoint + '/' + rowId +'?ngrestCallType=update&fields='+fieldName, angular.toJson(json, true), {
+				var pk = $scope.getRowPrimaryValue(value);
+				$http.put($scope.config.apiEndpoint + '/' + pk +'?ngrestCallType=update&fields='+fieldName, angular.toJson(json, true), {
 					  ignoreLoadingBar: true
 				});
 			});
@@ -435,6 +467,17 @@
 		};
 
 		$scope.getRowPrimaryValue = function(row) {
+			var pk = $scope.config.pk;
+
+			if (angular.isArray(pk)) {
+				var values = [];
+				angular.forEach(pk, function(name) {
+					values.push(row[name]);
+				});
+
+				return values.join();
+			}
+
 			return row[$scope.config.pk];
 		};